# server/openapi.yaml
openapi: 3.0.3
info:
  title: Valance API
  version: 2.0.0
  description: |
    Social habit-tracking backend. Base URL: http://localhost:3000/api/v1

    **Auth:** Bearer token (Firebase JWT) in Authorization header. Dev mode: X-Dev-User-Id header bypasses auth.

    **Response format:** `{ status: "ok", data: ... }` or `{ status: "error", error: { code, message } }`

    **Dates:** YYYY-MM-DD strings. Timestamps: ISO 8601.

    **Constants:**
    - XP/Sparks: light=5, moderate=10, intense=20, perfect_day=25, welcome=50
    - Streak milestones: 7d=+30, 30d=+100, 100d=+500
    - Ranks: bronze(0), silver(500), gold(2000), platinum(5000), diamond(15000)
    - Group tiers: spark(0), ember(7), flame(21), blaze(66)
    - Group chain: 100%=gold, ≥75%=silver, <75%=broken
    - Max habits: 7 active per user
    - Freeze: 100 Sparks, 1/group/day
    - Nudge limit: 5/day, 1/pair/day
servers:
  - url: http://localhost:3000/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /shop/items:
    get:
      tags: [Shop]
      summary: List shop items
      description: Returns all items with `owned` (boolean) and `meets_rank` (boolean) per item for the authenticated user.
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [theme, flame, animation, card_style, font, pattern, icon, profile_frame, profile_banner, name_color, celebration, milestone_card, party_badge, sound, freeze_animation, summary_style, party_entrance, qol, seasonal]
      responses:
        '200':
          description: Shop items
          content:
            application/json:
              example:
                status: ok
                data:
                  - id: "neon-terminal"
                    name: "Neon Terminal"
                    category: "theme"
                    sparksCost: 200
                    minRank: "silver"
                    owned: false
                    meets_rank: true
  /shop/purchase/{itemId}:
    post:
      tags: [Shop]
      summary: Purchase item
      description: |
        Validates: sufficient Sparks, meets min rank, not already owned, seasonal availability.
        Deducts Sparks atomically.
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Purchased
          content:
            application/json:
              example:
                status: ok
                data:
                  item: { id: "flame-blue", name: "Blue Flame", sparksCost: 50 }
                  remaining_sparks: 510
        '400':
          description: Insufficient Sparks or rank too low
        '409':
          description: Already owned
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase JWT token. Dev mode alternative — X-Dev-User-Id header.
...